Extract ensureFiles to separate file.
