+++ TestFoundation/TestNSString.swift
@@ -495,33 +495,6 @@ class TestNSString : XCTestCase {
     	return lhs.compare(rhs.bridge(), options: .CaseInsensitiveSearch) == .OrderedSame
     }
 
-    private func ensureFiles(fileNames: [String]) -> Bool {
-        var result = true
-        let fm = NSFileManager.defaultManager()
-        for name in fileNames {
-            guard !fm.fileExistsAtPath(name) else {
-                continue
-            }
-            
-            var isDir: ObjCBool = false
-            let dir = name.bridge().stringByDeletingLastPathComponent
-            if !fm.fileExistsAtPath(dir, isDirectory: &isDir) {
-                do {
-                    try fm.createDirectoryAtPath(dir, withIntermediateDirectories: true, attributes: nil)
-                } catch let err {
-                    print(err)
-                    return false
-                }
-            } else if !isDir {
-                return false
-            }
-            
-            
-            result = result && fm.createFileAtPath(name, contents: nil, attributes: nil)
-        }
-        return result
-    }
-
     func test_stringByTrimmingCharactersInSet() {
         let characterSet = NSCharacterSet.whitespaceCharacterSet()
         let string: NSString = " abc   "
